Update persona.js
No se puede borrar a la persona si tiene libros asociados.

+++ src/api/persona.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const PersonaModel = require('../models/persona');
+const LibroModel = require("../models/libro");
 
 router.post('/', async(req, res, next) => {
     const persona = new PersonaModel({
@@ -59,16 +60,29 @@ router.put('/:id', async(req, res) => {
     }
 });
 
-router.delete('/:id', async(req, res, next) => {
-    const { id } = req.params;
-    try {
-        const personaBorrada = await PersonaModel.findByIdAndDelete(id);
-        res.status(200).json(personaBorrada);
-        console.log('Se borro correctamente');
-    } catch (error) {
-        res.status(413).send(error, { mensaje: 'Error inesperado, No existe esa persona, Esa persona tiene libros asociados no se puede eliminar' });
-        next(error);
+router.delete("/:id", async (req, res, next) => {
+  const { id } = req.params;
+  try {
+    const persona = await PersonaModel.findById(id);
+    const libro = await LibroModel.find({ persona_id: persona.id });
+    if (libro == "") {
+      const personaBorrada = await PersonaModel.findByIdAndDelete(id);
+      console.log("Se borro correctamente");
+      res.status(200).json(personaBorrada);
     }
+    res
+      .status(200)
+      .send({
+        mensaje:
+          "No se puede borrar a la persona porque tiene libros prestados.",
+      });
+  } catch (error) {
+    res.status(413).send(error, {
+      mensaje:
+        "Error inesperado, No existe esa persona, Esa persona tiene libros asociados no se puede eliminar",
+    });
+    next(error);
+  }
 });
 
 module.exports = router;
